Move setup_logging into logs module
