+++ deafrica/data/create_mosaic.py
@@ -11,7 +11,7 @@ from odc.aws import s3_client, s3_dump, s3_head_object
 from pystac.asset import Asset
 from rio_stac import create_stac_item
 
-from deafrica.utils import setup_logging
+from deafrica.logs import setup_logging
 
 
 def _save_opinionated_cog(
