+++ deafrica/data/wapor_v3.py
@@ -23,6 +23,7 @@ from eodatasets3.stac import to_stac_item
 from odc.aws import s3_dump
 
 from deafrica.easi_assemble import EasiPrepare
+from deafrica.logs import setup_logging
 from deafrica.utils import (
     check_directory_exists,
     check_file_exists,
@@ -35,7 +36,6 @@ from deafrica.utils import (
     is_s3_path,
     is_url,
     odc_uuid,
-    setup_logging,
 )
 
 # Set log level to info
