+++ deafrica/monitoring/s2_c1_gap_filler.py
@@ -14,11 +14,11 @@ from odc.aws.queue import get_queue, publish_messages
 from rasterio.session import AWSSession
 
 from deafrica import __version__
+from deafrica.logs import setup_logging
 from deafrica.utils import (
     find_latest_report,
     read_report_missing_scenes,
     send_slack_notification,
-    setup_logging,
     slack_url,
     split_list_equally,
 )
