+++ deafrica/monitoring/s2_c1_gap_report.py
@@ -10,10 +10,10 @@ from odc.aws import s3_client, s3_dump
 from urlpath import URL
 
 from deafrica import __version__
+from deafrica.logs import setup_logging
 from deafrica.utils import (
     list_inventory,
     send_slack_notification,
-    setup_logging,
     slack_url,
     update_stac,
 )
