+++ deafrica/monitoring/s2_gap_filler.py
@@ -17,12 +17,12 @@ from rasterio.session import AWSSession
 from shapely import geometry
 
 from deafrica import __version__
+from deafrica.logs import setup_logging
 from deafrica.utils import (
     find_latest_report,
     limit,
     read_report_missing_scenes,
     send_slack_notification,
-    setup_logging,
     slack_url,
     split_list_equally,
 )
