+++ deafrica/monitoring/s2_gap_report.py
@@ -11,9 +11,9 @@ from odc.aws.inventory import list_inventory
 from urlpath import URL
 
 from deafrica import __version__
+from deafrica.logs import setup_logging
 from deafrica.utils import (
     send_slack_notification,
-    setup_logging,
     slack_url,
     update_stac,
 )
