+++ deafrica/platform/warn_unused_sandbox_volumes.py
@@ -4,7 +4,7 @@ from datetime import datetime, timedelta
 import boto3
 import click as click
 
-from deafrica.utils import setup_logging
+from deafrica.logs import setup_logging
 
 log = setup_logging()
